% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{analysis_control}
\alias{analysis_control}
\title{Specify control parameters}
\usage{
analysis_control(
  crossfit_folds = 5L,
  mlr3superlearner_folds = 10L,
  mlr3superlearner_discrete = FALSE
)
}
\arguments{
\item{crossfit_folds}{[\code{numeric(1)}]\cr The number of crossfit folds.}

\item{mlr3superlearner_folds}{[\code{numeric(1)}]\cr The number of `mlr3superlearner` folds.}
}
\value{
A list of control parameters
}
\description{
Specify control parameters
}
\examples{
analysis_control(crossfit_folds = 5)
}
