% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incremental.R
\name{incremental.inclusion}
\alias{incremental.inclusion}
\title{Subgroup analysis based on incremental inclusion}
\usage{
incremental.inclusion(
  data,
  subgroup = "R",
  treatment = "tt",
  outcome = "Y",
  covariates,
  deltas = c(1.5, 2, 5),
  learners,
  control = analysis_control(crossfit_folds = 5)
)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
A \code{data.frame} containing all necessary variables
for the analysis.}

\item{subgroup}{[\code{character(1)}]\cr
The column name of the subgroup status variable.}

\item{treatment}{[\code{character(1)}]\cr
The column name of the treatment variable.}

\item{outcome}{[\code{character(1)}]\cr
The column name of the outcome variable.}

\item{covariates}{[\code{character}]\cr
A character vector containing the column names of baseline covariates to be included.}

\item{deltas}{[\code{numeric}]\cr
A sequence of positive numbers specifying the increments to the odds of the focal subgroup.}

\item{learners}{[\code{character}]\cr
A vector of \code{mlr3superlearner} algorithms
for estimation of the outcome regressions. Default is \code{"glm"}, a main effects GLM.}

\item{control}{[\code{analysis_control}]\cr
Control parameters for the estimation procedure. Use \code{analysis_control()} to set these values.}
}
\value{
A \code{list} object containing a \code{data.frame} named "ATE_deltas" with the following columns:

\item{delta}{The increment odds ratio (delta) values specified by the user, for which the incremental subgroup effects are estimated.}
\item{estimand}{The incremental subgroup effect "ATE|Rq_delta{delta_value}" and the potential outcomes  "Y1|Rq_delta{delta_value}" and "Y0|Rq_delta{delta_value}", where "delta_value" is replaced with each of the user-specified delta values (rounded to 2 decimal places).}
\item{estimate}{The point estimate.}
\item{std.error}{The standard error estimate.}
\item{ci.low}{The lower bound of the 0.95 confidence interval.}
\item{ci.high}{The upper bound of the 0.95 confidence interval.}
}
\description{
Subgroup analysis based on incremental inclusion
}
\examples{

library(origami)
library(tidyverse)
library(glue)
library(mlr3superlearner)
library(mlr3extralearners)

data(data)
# specify model estimation methods
learners <- c("glm", "ranger", "gam")

# specify the increment odds ratio delta values
deltas <- c(1/10, 1/5, 1, 5, 10)

# run
out <- Subgroup::incremental.inclusion(
  data = data,
  subgroup = "R",
  treatment = "tt",
  outcome = "Y",
  covariates = grep("^X", colnames(data), value = TRUE),
  deltas = deltas,
  learners = learners
)
out$ATE_deltas

}
