\name{subgroup}
\alias{subgroup}
\docType{package}
\title{
subgroup Package
 - Algorithms and Tools for Efficient Subgroup Discovery
}
\description{
The subgroup discovery package contains a set of efficient and effective
 tools and algorithms for subgroup discovery.
}
\details{
\tabular{ll}{
Package: \tab subgroup\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2014-08-01\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Martin Atzmueller

Maintainer: Martin Atzmueller <martin@atzmueller.net>
}
\references{
  \enumerate{
	\item Martin Atzmueller and Frank Puppe. SD-Map - A Fast Algorithm for Exhaustive Subgroup Discovery. Knowledge Discovery in Databases: PKDD 2006, LNAI 4213, pp. 6-17, Springer Verlag, 2006.
	\item Martin Atzmueller and Florian Lemmerich. Fast Subgroup Discovery for Continuous Target Concepts. In: Foundations of Intelligent Systems, LNCS 5722, pp. 35-44, Springer Verlag, 2009.
	\item Florian Lemmerich and Mathias Rohlfs and Martin Atzmueller. Fast Discovery of Relevant Subgroup Patterns. In: Proc. 23rd FLAIRS Conference, AAAI Press, 2010.
  }
}
\keyword{ package }

